support cloneRangeIfSame on read-only subvolumes (requires CAP_SYS_ADMIN)
The BTRFS_IOC_FILE_EXTENT_SAME ioctl requires that the destistation
files have been opened for writing, unless the process has the
CAP_SYS_ADMIN capability. The problem with opening them for writing is
that it fails on read-only subvolumes. To address this issue, we detect
if the process has the CAP_SYS_ADMIN capability and choose between read
or write mode based on that. This allows us to support deduplication on
read-only subvolumes, but for privileged users only.